refactor progress... breaking allthe transforms methods out in to their own module_function modules
